Adding new tests and refactoring
