+++ tests/behavioural/features/test_sign_in_page.feature
@@ -11,8 +11,20 @@ Feature: Open the sign in page
     When I click sign-out
     Then I should return to the sign in page
 
-  Scenario: User signs in with incorrect credentials
+  Scenario: User signs in with incorrect email
     Given I go to the sign-in page
-    And I enter the incorrect credentials
+    And I enter the incorrect email
+    When I click sign-in
+    Then I should receive a sign-in error
+
+  Scenario: User signs in with incorrect password
+    Given I go to the sign-in page
+    And I enter the incorrect email
+    When I click sign-in
+    Then I should receive a sign-in error
+
+  Scenario: User signs in with incorrect email and password
+    Given I go to the sign-in page
+    And I enter the incorrect email
     When I click sign-in
     Then I should receive a sign-in error
